+++ SellFlow/Model/AnuncioSituacaoModel.cs
@@ -7,9 +7,10 @@ namespace SellFlow.Model
 {
     public class AnuncioSituacaoModel
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string nome { get; set; }
         public string descricao { get; set; }
         public bool ativo { get; set; }
+
     }
 }
